Fix negative encoder range
So if you install your encoder on the "other" side, you need
to configure it as a negative resolution.  The old code handled
this very badly and effectively returned really bad results if
you configured a negative resolution.

We also write a bunch of unit tests to validate the values are now
correct.