Don't double negative
So when an encoder resolution is negative, I was reversing the
result returned to the DSC.  This is *WRONG*.  The DSC is supposed to
do the needful when the resolution is negative.  This caused a problem
(usually for Alt encoders) whereby up was down and down was up when the
resolution is negative.

We now cacluate the value to return to the DSC the same regardless of
if the resolution is positive or negative and let the DSC do the right
thing.

fixes: #46